Perform screenshots after CA commits
It can happen CA has outstanding transactions while we attempt to
get a screenshot. That could result in errors like
`'NSInternalInconsistencyException', reason: 'unexpected start state'`
which in turn causes apps to exit.

To avoid this situation, this change schedules the screenshot to
happen in a safe moment after CA transactions are committed.